Rename rule 'INT' to 'ddigits' as in EGADSS Arden Compiler Grammar
This fixes the error: "Tried to register a value converter for a
fragment terminal rule: INT".
As we don't use the org.eclipse.xtext.common.Terminals for our grammar,
we can use the rule 'INT', even as a fragment, but xtext always seems to
try to register a value converter to the 'INT' rule, which is not
possible for fragments.