Update to Xtext v2.10 + Directory structure update
This cleans up the project/directory structure by splitting the editor
into two features and an update-site:
- arden.plugin.editor: Xtext editor stuff
- arden.plugin.compiler: Arden2ByteCode Integration (run/debug)
- arden.plugin.update-site: Update site for both features

The compiler needs to be added to the arden.plugin.compiler project as
arden2bytecode.jar.
Previously this used a "linked resource" and you had to unzip the
contents of the jewelcli.jar into the project.

This also cleans up the MANIFEST.MF files