Fix errors on some subtractions
The lexer is greedy and tries to invoke the ISO_DATE terminal rule as
soon as it sees four digits and a "-". 
This leads to an error when the subsequent characters don't fit the
ISO_DATE terminal rule and  the lexer can't fall back to an alternative
terminal rule.
This is a simple fix, which offers an alternative terminal rule to fall
back to.