Update to v1.2.0
This also changes the site to a simple repository, instead of a composite repository.
Update-sites usually only contain the newest version of features and plugins, so the old versions are removed.
Note: If old versions should remain available (e.g. if a newer version doesn't support an older eclipse version, but the old eclipse version still needs to be supported), they can be kept in seperate update-sites.