Make all keywords uppercase
Keywords in the examples given by the specification are usually
uppercase. 
This is similar to SQL, where keyword are preferably in uppercase to
distinguish them from identifiers.
Arden Syntax is often written in plain text editors without syntax
highlighting, so I think it would be best to make uppercase the default
style.
Lowercase is still supported though.